refactor: move brand to /constants
