--- app/lib/brand/config.ts
@@ -1,14 +1,0 @@
-export const siteConfig = {
-  siteTitle: 'Remix SaaSkit',
-  siteDescription: 'A starter kit for building SaaS applications with Remix.',
-  siteUrl: 'https://remix-saaskit.com',
-  siteImage: '/images/og-image.png',
-  twitterHandle: '@remix_saas',
-  facebookHandle: 'remix.saas',
-  linkedinHandle: 'remix.saas',
-  instagramHandle: 'remix-saas',
-  email: 'hello@remix-saas.run',
-  phone: '',
-  address: '',
-  favicon: '/favicon.ico',
-}
