+++ app/routes/auth+/login.tsx
@@ -18,11 +18,11 @@ import { getSession, commitSession } from '#app/modules/auth/auth-session.server
 import { validateCSRF } from '#app/utils/csrf.server'
 import { checkHoneypot } from '#app/utils/honeypot.server'
 import { useIsPending } from '#app/utils/misc'
+import { siteConfig } from '#app/utils/constants/brand'
 import { Input } from '#app/components/ui/input'
 import { Button } from '#app/components/ui/button'
 import { ROUTE_PATH as DASHBOARD_PATH } from '#app/routes/dashboard+/_layout'
 import { ROUTE_PATH as AUTH_VERIFY_PATH } from '#app/routes/auth+/verify'
-import { siteConfig } from '#app/lib/brand/config'
 
 export const ROUTE_PATH = '/auth/login' as const
