+++ app/routes/dashboard+/_index.tsx
@@ -5,8 +5,8 @@ import { useTranslation } from 'react-i18next'
 import { requireUser } from '#app/modules/auth/auth.server'
 import { prisma } from '#app/utils/db.server'
 import { cn } from '#app/utils/misc.js'
+import { siteConfig } from '#app/utils/constants/brand'
 import { buttonVariants } from '#app/components/ui/button'
-import { siteConfig } from '#app/lib/brand/config'
 
 export const meta: MetaFunction = () => {
   return [{ title: `${siteConfig.siteTitle} - Dashboard` }]
