+++ app/routes/dashboard+/checkout.tsx
@@ -7,9 +7,9 @@ import { requireSessionUser } from '#app/modules/auth/auth.server'
 import { PLANS } from '#app/modules/stripe/plans'
 import { prisma } from '#app/utils/db.server'
 import { useInterval } from '#app/utils/hooks/use-interval'
+import { siteConfig } from '#app/utils/constants/brand'
 import { ROUTE_PATH as DASHBOARD_PATH } from '#app/routes/dashboard+/_layout'
 import { buttonVariants } from '#app/components/ui/button'
-import { siteConfig } from '#app/lib/brand/config'
 
 export const ROUTE_PATH = '/dashboard/checkout'
